moved user dtos
