+++ src/Infrastructure/Controller/UserController.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Infrastructure\Controller;
 
-use App\Domain\User\Dto\User as UserDto;
+use App\Infrastructure\Dto\User\User as UserDto;
 use App\Domain\User\Entity\User;
 use Nelmio\ApiDocBundle\Annotation\Model;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
