+++ src/Infrastructure/Dto/User/ResetPassword.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\User\Dto;
+namespace App\Infrastructure\Dto\User;
 
 use Symfony\Component\Validator\Constraints as Assert;
